Abstracted Graphic API to ease future integration of other API (DIRECT3D11 for instance)

+++ comet/include/comet/staticMesh.h
@@ -3,7 +3,7 @@
 #include <vector>
 #include <memory>
 #include <comet/vertex.h>
-#include <comet/vertexBufferLayout.h>
+#include <rendering/vertexBufferLayout.h>
 
 namespace comet
 {
@@ -21,14 +21,14 @@ namespace comet
         const std::vector<Vertex>& getVertices() const { return m_vertices; }
         std::vector<Vertex>& getVertices() { return m_vertices; }
         uint32_t getVertexCount() const { return m_vertexCount; }
-        uint32_t getVerticesSize() const { return m_vertexCount * sizeof(Vertex); }
+        size_t getVerticesSize() const { return m_vertexCount * sizeof(Vertex); }
 
         bool hasIndices() const { return m_indexCount != 0; }
         void setIndices(const uint32_t* indices, uint32_t indexCount);
         const std::vector<uint32_t>& getIndices() const { return m_indices; }
         std::vector<uint32_t>& getIndices() { return m_indices; }
         uint32_t getIndexCount() const { return m_indexCount; }
-        uint32_t getIndicesSize() const { return m_indexCount * sizeof(uint32_t); }
+        size_t getIndicesSize() const { return m_indexCount * sizeof(uint32_t); }
 
     private:
         uint32_t m_indexCount{0};
